New Coding Guidelines
Hi friends :)

Since his topic came up with both of you this week, I went ahead and wrote up a first draft of coding guidelines for our React code.

The goal is to discuss and agree on a set of rules to keep our codebase clean and easy to understand for everyone.

The baseline of these rules come form the rules we used at FB where we spent a lot of time figuring out how to make code review easy and effective.

So you'll see that these guidelines will result in a lot of small files with clear rules on when to break things out and where to put them. This has helped to get an overview of what is available and how the code is structured just from looking at the file structure.

Let me know what you think in the comments, and we can also discuss in person on Friday :)